Read sentinel raw dn band data as float64
As all other operations work with float64,
the dn values are also read as float64 (they are actually uint16, but just to make sure all the data has a similar dtype). This should not be a big deal as even 32 bits are enough for sentinel images as far as I managed to research.

New version 7.0.3

+++ sentinel_toolkit/colorimetry/sentinel_values.py
@@ -153,11 +153,11 @@ def dn_to_sentinel(raw_band_data, nodata_value, boa_offset, quantification_value
 
     Parameters
     ----------
-    nodata_value : float
+    nodata_value : int
         The nodata pixel value.
     raw_band_data : ndarray
         The raw sentinel-2 band(s) pixel values.
-    boa_offset : float scalar or ndarray
+    boa_offset : int scalar or ndarray
         The boa offset.
     solar_irradiance : float scalar or ndarray
         The solar irradiance.
